Use with keyword for opening target file
Simply reading a file is dangerous, it might happens that the file
might be open but not closed will leads to ResourceWarning: unclosed
file error, It is necessary to use with keyword which will close the
file first and then allow it to open it.

Changing r mode to rb mode as in Python3, While r mode is used then
it reads the file based on file encoding provided but with rb mode
it reads the stream as a binary which avoides the encoding issue.

Change-Id: Ia48c1cb0e436ac2c53fc6bd46c4a906c70a51325