Run hostname check for standalone
Run a hostname preflight check when running the standalone as this can
cause issues with rabbit/mysql/pacemaker/puppet. The standalone check is
not run when an undercloud is being deployed or if output only is
specified. The undercloud has a different set of preflight checks run
prior to deployment so we do not need to rerun the hostname check when
the undercloud is being deployed.

Conflicts:
	tripleoclient/tests/test_utils.py
	tripleoclient/tests/v1/tripleo/test_tripleo_deploy.py
	tripleoclient/utils.py
	tripleoclient/v1/tripleo_deploy.py
	tripleoclient/v1/undercloud_preflight.py

Depends-On: https://review.openstack.org/#/c/637309/
Change-Id: I4de51d792f004575e0e802a1d340478d6ae9f848
Closes-Bug: #1814564
(cherry picked from commit 63f4175e768796970801a36dee9585b3d89a9531)