Allow 'container image prepare' to support drivers
This updates the 'overcloud container image prepare' command
so that it sets the neutron_driver variable correctly. This
allows tripleo-common to detect if OVN or ODL drivers are in
used and prepares the resulting output files for container
images and parameters accordingly.

Previously you could only access this functionality if you
used the new ContainerImagePrepare workflow (requires Mistral).

Change-Id: Icb5b20bfd98a704f3f33f739bc92f77550b52c26
Closes-bug: #1816663