Ensure created symlink is in $PATH
With "sudo", the $PATH is overriden an really basic:
PATH=/sbin:/bin:/usr/sbin:/usr/bin
It doesn't list the /usr/local/(s)bin, hence the created
symlink wasn't seen by the script.

It also push that symlink creation in the __new__ special
method in order to make it cleaner.

Finaly, it takes care of lint/pep8 failures, not related to
this change (how is that even possible?!)

Co-Authored-By: Natal Ngétal <hobbestigrou@erakis.eu>
Closes-Bug: 1817365
Partial-Bug: 1816446

Change-Id: If09eb32f43b2c26c8cd8a89f2e862db1dd91dfed