Fix ansible-playbook-3 symlink
The symlink needs to be created with sudo because if validations are
enabled, the call to run ansible-playbook-3 is before the "sudo tripleo deploy
command".

This patch moves the symlink into a function, that will be called before
the preflight if enabled and during the tripleo deploy/upgrade command,
which should cover standalone deployment and upgrade.

Change-Id: I1ce9b3df6c937cd796728d4fc5921fcf1023e5db