Container image config file for builds optional
Make the config option optional for the container image build process.
In previous releases we have shipped a basic yaml file that contains the
list of container to build. We have moved to generating using a jinja
file instead which we ship so we don't need to manage this static file
going forward.

Change-Id: Id689c54f393affe103082e329efaf4ebf376e397
Depends-On: https://review.openstack.org/#/c/639846/