[Queen Only] Add force option to undercloud update.
When the overcloud is in a failed state we can't run the undercloud
upgrade command anymore. But we may need to inject code in mistral for
instance to get out of the overcloud failure.

The force option enables the user to get out of this loop.

Queen only as instack undercloud is removed from Rocky on.

Change-Id: I9e7fe8cfd96262938d463de0d6c45ccedc9989be
Closes-Bug: #1806683
Depends-On: https://review.openstack.org/622290