Add minion installation
Add the ability to install an undercloud minion which is connected to an
original undercloud. This minion can have either heat-engine or
ironic-conductor deployed on it.

Adds two new openstack commands for the minion install and a new
minion.conf can be used to configure them.

openstack undercloud minion install
openstack undercloud minion upgrade

Depends-On: https://review.opendev.org/#/c/656984
Change-Id: I61832f5088be172eaf31b36a9cca8dc289580bb2
Related-Blueprint: undercloud-minion