Send terminal erase sequence on tripleo deploy
We're seeing some escape sequences being dropped when we run the
deployment via the undercloud. This causes a user to have to run a reset
after the undercloud installation in order to be able to use their shell
command line again.  We can avoid this if we send an terminal erase line
sequence to trim the rest of the line at the end of the deploy.

Conflicts:
	tripleoclient/tests/v1/tripleo/test_tripleo_deploy.py

Change-Id: Ibb74bfebfaab2cd122abe1bc3cd5bbd2562ccfc5
Closes-Bug: #1833302
(cherry picked from commit e1a051d6ac890cd68f5bf5fff251fa445d7ad8f9)