Allow running validations with custom extra variables
This patch adds the ability to pass extra variables to the validations
execution through Mistral and 'ansible-playbook'.

By default, the CLI is running the validations through
'ansible-playbook', then the operator will have two possible ways to
process it;

Either passing a Dict with the new --extra-vars argument

```
$ openstack tripleo validator run \
    --extra-vars '{"min_undercloud_ram_gb": 2}'
    --validation-name undercloud-ram
```

or passing the absolute path of a JSON/YAML file using --extra-vars-file
argument.

```
$ cat /home/stack/vars.json
{
  "volumes": [
    {"min_size": 10, "mount": "/var/lib/docker"},
    {"min_size": 3,  "mount": "/var/lib/config-data"},
    {"min_size": 3,  "mount": "/var/log"},
    {"min_size": 5,  "mount": "/usr"},
    {"min_size": 20, "mount": "/var"},
    {"min_size": 25, "mount": "/"}
  ]
}
$ openstack tripleo validator run \
    --extra-vars-file /home/stack/vars.json
    --validation-name undercloud-disk-space
```

If the operator wants to use Mistral, she/he will just need to add the
--use-mistral argument to the command line in the previous
examples.

Note that Mistral will only support a valid JSON file as input.

Change-Id: Idff6bcebc041db483d64396249b040e6bdd778bd
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>