Add --parameters and --create-vars-file arguments to the list subcommand
The 'openstack tripleo validator list' subcommand can now get only the
available parameters for the validations using the new --parameters
argument.

```
$ openstack tripleo validator list \
    --parameters                   \
    [--validation-name <validation_id>[,<validation_id>,...] |
     --group <group>[,<group>,...]]
```

Here is an output example:

```
Waiting for messages on queue 'tripleo' with no timeout.
{
    "undercloud-cpu": {
        "parameters": {
            "min_undercloud_cpu_count": 8
        }
    },
    "undercloud-ram": {
        "parameters": {
            "min_undercloud_ram_gb": 24
        }
    }
}
```

The --create-vars-file allow the operator to generate either a JSON or a
YAML file containing only the parameters of one or multiple validations.
This file will be available to pass as extra vars to the validations
execution.

```
$ openstack tripleo validator list \
    --parameters                   \
    --create-vars-file [json|yaml] /home/stack/myvars \
    [--validation-name <validation_id>[,<validation_id>,...] |
     --group <group>[,<group>,...]]
```

Change-Id: I6e2255c0d490ee8105f0757d02f5d8fba1d4fa20
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>