Output oslo messaging info for the minion
The oslo messaging host configuration does not actually use the endpoint
map so we need to be able to provide this information as part of the
output yaml file for use with a minion. This code grabs the ctlplane
ip out of the rolenetmap for the undercloud and uses it to construct the
oslo messaging node names value that is used when configuring the service.

Blueprint: undercloud-minion
Change-Id: I5dc103452f3e7484ea81df15d75b606d2c74859a