Run Validations with ThreadPoolExecutor
The validations are sequentially executed through ansible-playbook and
could take some time to complete. This patch adds support of running all
the validations in parallel through a ThreadPoolExecutor in order to
save processing time.

Without this patch, it takes almost ~5min to run all the validations but
only ~1min10 with the ThreadPoolExecutor.

- Add --worker/-w argument to give the maximum number of threads that
  can be used to execute the given validations
- Python 2.7: use futures backport instead, ThreadPoolExecutor is default
- Use six in tripleo_validator.py for python 3 compatibility

Change-Id: Ia805a556bc26700a3eb520ed72e90b37546901b8
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>