Pass undercloud's private key to inventory used in package_update.
As the Ansible playbook is being executed from the mistral_executor
container, we need to establish a conection to the Undercloud via ssh
not local as it was being done up to now. In order to be able to ssh
into the host we need to provide with the tripleo-ansible's user ssh
key.

Depends-On: https://review.opendev.org/#/c/676415/

Closes-Bug: #1832932
Change-Id: I20948d49b3023dc76acff8779bbe182886e55838