Use reset to fix cmdline
Reset is an application provided via ncurses which is useful in
resetting a terminal when things go wrong. Since podman is eating some
of the terminal output, it can leave the cmdline in a broken state at
the end of the undercloud install. We tried to address it outselves but
let's reuse an existing application instead. -I being passed to reset
skips the initialization strings which can erase the output.

Conflicts:
	tripleoclient/tests/v1/tripleo/test_tripleo_deploy.py

Change-Id: Id9aa7910032588605977f4725385503cf0a235bc
Closes-Bug: #1833302
(cherry picked from commit fddcb9e0755beec8eff3b7246a3b7b799a8ec7f2)