Clarify the purpose of --http-boot
In a multiarch cloud when using "openstack overcloud image upload" the
argument --http-boot is of vital importance. This argument must be
customized such that a distinct folder is created for the images of
each arch (and possibly platform). Now, reflect this in docstring /
help text (tripleo-docs already has some content, luckily).

Change-Id: I707ed1de6aa0347185769baa2ee0539bcb6d2df4
Closes-Bug: #1840711