Disable inflight validations by default
The inflight validations caused timeouts in the TripleO CI.
They should not be enabled by default. New features should not be
enabled by default in general, they should be optional.

So this patch changes the option to be --inflight-validations,
and make sure the inflights are disabled by default everywhere it's
called.

Change-Id: I082628a9480686d6a7801056c3b4bf332b4e3d95
Related-Bug: #1844446