Don't use --registry-url for image show auth
This doesn't work for registries like docker.io because they expect
the authentication url to include the namespace so they can determine
the allowed scope of the request.

Instead the image reference is used for authentication, and the docs
include examples to make it clear that the registry host needs to be
included in the image reference.

This change also makes json the default formatter for the show
command, so it can be used as a drop-in replacement for "skopeo
inspect" and so that the output format is consistent with the docker
commands.

Change-Id: I30ce5c53e3eac758f52f31cc9ed3f2dff35eb4cb
Closes-Bug: #1845400