Handle a deployment that was manually cancelled.
When doing a deployment or update, if the end-user uses CTRL+c
(SIGINT) to end it early, it will fail complaining that there is
no passwords key in the plan. We will now handle this and cleanup
the existing plan.

Change-Id: Ie7f1a4269936f98ad86db4993992e0b9a27edf73
Closes-Bug: #1819672
Resolves: rhbz#1753251
Signed-off-by: ekultails <ekultails@gmail.com>
(cherry picked from commit 72850060e058274bb55aadbf2963f4f46ba3be3a)