Undercloud hosts entries in overcloud parameters
When running overcloud deploy set the UndercloudHostsEntries
parameter to include the <shortname>.ctlplane entry from
/etc/hosts on the undercloud.

With IPv6 the overcloud nodes cannot pull images from the
undercloud using an IPv6 address. By adding the host entry
for the undercloud in the overcloud nodes hosts file a name
can be used instead of the IPv6 address.

Related-Bug: #1836578
Change-Id: Id9271460b20e602f935f1eddd70f989cd605b1f4