Support undercloud fake-hardware ironic driver
When including fake-hardware in the undercloud.conf
enabled_hardware_types, some of the fake-hardware default interfaces
are not enabled so it can't be used.

This change allows nodes to be created with driver fake-hardware. If
this proves useful for development or some scale testing then it might
be desirable to add fake-hardware by default to the
enabled_hardware_types list.

Change-Id: I1a6b4df073b39924e6855ea865cfc856a1bdfe6b