Write minimal paste configuration for heat-all.
Instead of relying on the system's paste configuration for heat api, we
can write a minimal configuration that is needed by the heat-all
process. This works in a similar way that we generate a minimal
configuration for heat.conf.

It also ensures that the configuration can be reprodcued if needed
without having to rely on what may have been configured on an individual
system.

Change-Id: I27e98fa77966213fd14e9a5e3187fd4e02db0a10