Add tripleo launch heat command
Adds a new command under the tripleo namespace for launching the all in
one heat-all process in the foreground. The new command reuses the code
from heat_launcher.py already in place for the tripleo deploy commands.

Launching the heat-all process separately from tripleo deploy allows for
easily spinning up a heat-all environment for stack creation and
generating the config-download ansible content for development and test.

Change-Id: I16d23faca6d8c2d0abf13dbdf91b810717ac585e