Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value. This change
showed we needed a newer version of requests in lower constraints,
aliging with the version required by tripleo-common. This change showed
we needed a newer version of openstacksdk in lower constraints, aliging
with the version required by metalsmith. This change showed we needed a
newer version of keystoneauth1 in lower constraints, aliging with the
version required by openstacksdk. This change showed we needed a newer
version of decorator in lower constraints, aliging with the version
required by openstacksdk. This change showed we needed a newer version
of dogpile.cache in lower constraints, aliging with version required by
openstacksdk.

Change-Id: I4668fe763239a8c6380e82e64d5c9aa0657bb628