Refactor image upload command
In preparation for supporting deploys on a glance-less undercloud,
this change refactors the upload command to move all glance
interacting logic into a glance adapter class. This will allow the
adapter abstraction to be used for an 'upload' adapter which only
copies files on the local disk.

This change also does the following:
- moves repeated update-or-upload logic to an adapter method, reducing
  the complexity of take_action
- cleans up the mock usage to consistently use the
  patch decorator - the current approach may be causing unit test
  failures in some environments
- removes the GlanceV1 implementation, since the V1 API was removed in
  Newton(!) this support is just a maintenance burden

Change-Id: Ie28dabaf07ed36adb3e8f07ca8c141f93a9d79cc
Blueprint: nova-less-deploy