Local file based image uploader
This change modifies the "openstack overcloud image upload" to add a
--local argument which will copy the files to /var/lib/ironic/images
instead of uploading them to glance. /var/lib/ironic is chosen as it
is bind-mounted into the ironic containers on the undercloud.

The on-disk layout is prefixed by the architecture (and optionally the
platform) for example:
/var/lib/ironic/images/x86_64/overcloud-full.qcow2
/var/lib/ironic/images/x86_64/overcloud-full.initrd
/var/lib/ironic/images/x86_64/overcloud-full.vmlinuz
/var/lib/ironic/images/power9-ppc64le/overcloud-full.qcow2

When glance is switched off this will become the default way of
preparing image files for deployment.

Blueprint: nova-less-deploy

Change-Id: I4e1b8a7effca21b5aa054a78ddeaae7338ca9d2c