openstack overcloud node delete --baremetal-deployment
This change integrates baremetal provisioning with the existing
"openstack overcloud node delete" command.

Instead of providing node ID(s), the --baremetal-deployment argument
will specify the provisioning yaml with instances set to
provisioned:false for the nodes to be deleted.

Internally the command needs to expand the roles and compare the
indexes of unprovisioned instances with the indexes of the heat
resources in the server resource group, using the match to specify the
heat resource physical_resource_id to pass to the scale down workflow,
automating what is done for deployed-server scale-down[1].

This is a medium-risk backport because overcloud node delete is
sometimes run, but the code-path is only used when the
--baremetal-deployment argument is passed (which is not documented in
stable/train)

Blueprint: nova-less-deploy
[1] https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/features/deployed_server.html#scaling-down

Change-Id: I3376378cdffbe69b8d14aed683deba185ba1e0f5
(cherry picked from commit bd9302d5523855442dacf88add600a2f67b3a318)