Pass container name to update_nodes workflow
The container name needs to be passed with the stack name value to the
update_nodes workflow so that the right config container is used during
the update.

Previously, this was not an issue since all update actions just
defaulted to overcloud-config even if the stack name wasn't overcloud.

However, with the multiple overclouds feature and the multi-stack
feature, we need to make sure the right config container is used that
matches the stack name.

Closes-Bug: #1854991
Depends-On: Id6b4a5913141fc558c40f2cfd29f5e0741544ab4
Change-Id: Ib25d8f8fbfde8fb13bd6518fd2383b7e3ae7abee
(cherry picked from commit 819dab257d5b671f3194792c111ece7b4ecaab39)