Fix ansible-runner when verbosity is > 0
This change updates the run_ansible_playbook method to ensure that we're not
attempting to re-open the ansible runner stdout property when executing a
task with a verbosity setting of > 0. When the property is read the tmp file
has already been cleaned up resulting in stacktrace. There's no value in
re-opening the property as stdout is already been presented to the user so
the offending code has been deleted.

Change-Id: I4ab300e8a2a70adb1f79334aee44fde31d2cfa70
Signed-off-by: Kevin Carter <kecarter@redhat.com>