Add raised exception when playbook fails
This change adds a raised exception when the return code from a given
playbook run is not 0. This will ensure that any playbook failure is
captured, raising an exception from within the executing method. Prior
to this change exception handling was expected to be done in the calling
method, however, doing that would create lots of unnecessary code
duplication.

Change-Id: Ic742ec4653eb45a66c0d5c86d3d0ff31947be5c4
Signed-off-by: Kevin Carter <kecarter@redhat.com>