Add kwarg to set an ansible.cfg file
This change implements the ability to set the file path for an ansible
configuration file. If no option is provide a default ansible.cfg file
will be generated in the artifact path. This default ansible.cfg file
will, at this time, contain one tunable which will improve CPU
availability by reducing the internal polling interval. While most of
the ansible configuration is performned though the use of environment
variables, not all options have an exposed environment variable or
playbook tunable. By having the ability to define an ansible
configuration file, or generate one when needed, we should be able
to accomodate any configuration required.

Change-Id: Ic2ae04daf81f19afa763c2f6b1c0ae5e6467324b
Signed-off-by: Kevin Carter <kecarter@redhat.com>