Remove mistral workflow to return the ssh private key
This change uses the local file system to return the ssh private key which is
known to be stored in the working directory or in the users home folder.
As the method searchs, it attempts to open the key file to ensure that the
calling user has access to the key. In the event of a failure, the updated
method will return None, which the calling methods expect.

Story: 2007212
Task: 38437

Change-Id: I9c0ae96787e8f361e20b2fe4c77f4bf6e873022e
Signed-off-by: Kevin Carter <kecarter@redhat.com>