Allow import to specify image location ipxe=false
When performing a 'openstack overcloud node import', we ignore the
--http-boot param as the path should be fixed for the general case
when iPXE is enabled.

However when deploying on ppc64le, we can't use iPXE, meaning we
need to be able to specify the image location.  This change allows
http_boot to be used if provided, but defaults to the most likely
case if it isn't.

Change-Id: Ia2212b1bd197a2930d435071c60b9aae0684d882
(cherry picked from commit 4e9c5bf73dfb917168650eb096516608e5dcc346)