Revert "Default to zero verbosity when using Ansible."
This reverts commit 90eee3abc1efca0166b6e5fb05998aa54abcbc05.

I'm reverting this because it's a regression.

I'm open to not reverting this one if we want to fix it. If the
intention is to backport then we should revert this and correct it
before backporting.  We shouldn't touch the verbosity for anything
except ansible items.  I'm uncertain if this silenced the heat output
for the upgrade tasks but it impacted the deployment output.

Change-Id: I1bc27ca55a86d8d51960750d36ba825e5197b75f
Closes-Bug: #1863633