Fix default image registry address lookup
By using a function as the default for the registry url variables, the
function is actually executed on every invocation of openstack client
which is less than ideal.  This change moves the default registry lookup
to only occur if the option is not passed in by the user. This limits
the lookup of the local registry to only occur when the correct
openstack tripleo container iamge commands are run.

Change-Id: Iace7be9dbdcf681c5f2f25f5810aa11e9da1853c
Closes-Bug: #1863679
(cherry picked from commit 7d3f274a3c660655072f11a2b6aedd63405df5f9)