Add handler for exception raising
This change adds a switch that will enable or disable raising an 
exception when a playbook executes. This will allow some methods to 
return the RC and status information when a playbook is run, even when 
there's a failure. The default behaviour is to raise an exception on 
failure, but when fail_on_rc is set to False the run_ansible_playbook
method will return the status and rc information regardless of any 
failures.

To ensure we're not raising an exception from the ansible runner library
its been changed to RuntimeError.

Change-Id: I3af652615b5227144256074c05170d148f19bc1d
Closes-Bug: #1859182
Signed-off-by: Kevin Carter <kecarter@redhat.com>