baremetal provision expose concurrency and timeout
These arguments are available in the workflow but are not exposed as
arguments. This change fixes that, allowing users to tune for their
environment.

The new ansible version of this command already has these arguments
(next change in this series)

Change-Id: I00b224ec7e2334afd2dfe549f36dbec2987a1be1
(cherry picked from commit 74d4b37cdea967b39ec832698ab18a50f5704d94)