Add new interface to run arbitrary playbooks from a plan
This change is the first step to providing a path to remove the mistral
"derived_params" workflows. Deployers will now be able to execute playbooks
from within a given plan by passing in the `playbook_parameters` key which
contains a dictionary of playbook file paths, and all options applied to a
given playbook.

Story: 2007419
Task: 39118

Change-Id: I26ea1c01e17513d001e311d68fd239d3ba8c8a0d
Signed-off-by: Kevin Carter <kecarter@redhat.com>