Ensure the undercloud output dir constant is consistent
By default the output directory is the executing users home folder,
however, because this is meant to be user configurable, we need
to modify our constants to respect the users configured information.

This change ensures that the constant UNDERCLOUD_OUTPUT_DIR is consistently
by pre-loading the undercloud configration file when present and pulling
out the `output_dir` option. When this file is not present or the option
is undefined, the constant will use the executing users home folder; the
execting users home folder is the original default.

Related-Bug: #1868619

Change-Id: I94257d1d5ceb2795f5d9596a6a1d5066d881c3b7
Signed-off-by: Kevin Carter <kecarter@redhat.com>