Set scp_if_ssh to True in ansible client configuration
To avoid the warning "sftp transfer mechanism failed" during Ansible
run (also reported here:
https://github.com/ansible/ansible/issues/22127); let's set scp_if_ssh
to True.

The default is False; which means it'll first try sftp (which fails
since we use pipeline) and fallback on scp. To avoid the warning, let's
set it to True directly, so we use scp.

Closes-Bug: #1869735

Change-Id: I5474b45f4bbe15780370a23d587dea7dd811c073
Signed-off-by: Kevin Carter <kecarter@redhat.com>