Get event marker before doing update
As we now have ansible playbook running the stack create/update
and would only finish when the stack has been *_COMPLETE/*_FAILED,
we need to get the event marker before doing the update, else there
would be no events listed for subsequent deploy (stack update).

Left the call to utils.wait_for_stack_ready() for the time being.
We can just get the events as the stack operation is finished or
move the event polling to the playbook.

Change-Id: I746f3bb135fac309bb7e780648e3e8216d76770e