Remove OrchestrationDeployAction call in ffwd upgrade prepare
This was added for ffwd upgrade from N->Q, where we moved to use
scipt hook rather than os-apply-config. This does not seem relevant
any more is it would be removed for both normal upgrade[1] and ffwd
upgraded N->Q deployments.

Also, it won't be needed for Q->T ffu.

[1] https://review.opendev.org/560022

Change-Id: I5c55938b1c313e4ebf24f3babeb377e87bf7cd30