Update necessary packages before undercloud upgarde.
This used to be a manual step in the tripleo process. You still have
to update python-tripleoclient and tripleo-common manually.  But:

 - "openstack-tripleo-heat-templates"
 - "openstack-tripleo-validations"
 - "tripleo-ansible"

can all be installed from this process as we spawn another new process
"openstack tripleo upgrade".  Thus all those packages will be
available for it.

Note that tripleo-common should still be installed before running that
command as some of its code is used by tripleoclient.

Eventually adding version dependencies from python-tripleoclient to those
other packages (or from tht which would stay a manual step) looks
challenging as it would be hard to track which version work with which
version.

Change-Id: Ic258f314074007e7e5fd16d87448ffb7a3447fc3
Closes-Bug: #1869776