Add the ability to load vars from files
This change will provide the ability to load extra vars from files, instead
of having to pass options through the CLI parser. By loading vars from file
we can ensure options are made more safe and better handled, especially in
cases when a given option may be massive, as is the case with
`parameter_defaults`.

> A new argument has been added to the ansible playbook runner which will
  allow us to pass options into the method that will be stored in an
  extravars file, which is then dynamically loaded by ansible-runner.
  Information on extravars files can be seen here: [0].
> A test has been added to exercise the new extravars file capability.

[0]: https://ansible-runner.readthedocs.io/en/latest/intro.html#env-extravars

Closes-Bug: #1871338
Change-Id: I9675e587abf3f07e91319a40620a8f4c67fbf97b
Signed-off-by: Kevin Carter <kecarter@redhat.com>